refactor: 엔티티 user, auth 파일 구조 변경
