+++ src/pages/user/userpage/UserBioSection.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { getBioApi } from "@/app/api/userActivityApi";
+import { getBioApi } from "@/entities/user/api/userActivityApi.ts";
 
 interface BioSectionProps {
   userId: number;
